CHROMIUM: iwl7000: mvm: load regdomain at INIT stage
We used to load the regdomain only in the load stage,
this caused the 'iw phy phy0 reg get' command to fail if we
booted a machine with wifi off.
Therefor we should load it in INIT stage already.

Change-Id: I7d8d5c895bc467efbf81ea055dde366ea01cced1
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
iwl7000-tree: 14b3d14f8934a9917becd9650777a1a2ebe90bfa