FROMLIST: driver: ahci: Allow disabling dlpm for all bios version
In ahci_broken_lpm(), dmi->driver_data is encoded as a date: if the
date of the current BIOS is older than driver_data, the LPM horkage is
not applied.
Allow dmi->driver_data is be empty in case a problem is not fixed in any
BIOS version.

(am from https://patchwork.ozlabs.org/project/linux-ide/patch/20210320034033.396854-2-gwendal@chromium.org/)

BUG=b:180139522
TEST=Check device DLPM is "max_performance".
Check that setting it to "min_power" fails, thus is can not be enabled
on the SSD.

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: I2955d91ab6a3f26d0e30bbfa3548f5d4fe35a6ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2776398
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Rajat Jain <rajatja@google.com>
Commit-Queue: Rajat Jain <rajatja@google.com>
Tested-by: Rajat Jain <rajatja@google.com>