CHROMIUM: iwl7000: skip first element in the WTAS ACPI table
By mistake we were considering the first element of the WTAS wifi
package as part of the data we want to rid, but that element is the wifi
package signature (always 0x07), so it should be skipped.

Change the code to read the data starting from element 1 instead.

Change-Id: I70636c04e37b2b57a5df3ce611511f62203d27a7
Signed-off-by: Abhishek Naik <abhishek.naik@intel.com>
iwl7000-tree: e800dacaa1ac382a971ac190d5014743eb3ce028