CHROMIUM: iwl7000: use length-limiting format instead of modifying region name
The TLV region data we receive in iwl_dbg_tlv_alloc_region() is
constant, but we were erroneusly casting them to non-const.  This was
fixed upstream in v5.12, but we still try to modify the data in our
code.

Instead of force-terminating the original string by adding a \0 at the
end, use the string length limit in the format we use.

Change-Id: I1f1a53db0504c63a54a3316114d63d805cfe620f
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 2d9588569ba35cfbd69aa8eefe6366731d92896d