CHROMIUM: iwl7000: always initialize len in iwl_uefi_get_reduced_power()
The len argument will only be left unset in
iwl_uefi_get_reduced_power() if there was an error.  If there was an
error, we set reduce_power_loaded to true and, if reduce_power_loaded
is true, iwl_trans_set_reduce_power() will not use len (because it
won't try to allocate the dma block).

But to make it safer and cleaner, initialize len to 0 at the
iwl_uefi_get_reduced_power(), so that it's set in all error
situations.

Change-Id: Icdcdab3ba5e3a47d3be3d2146a42ae4102c33980
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: dd6b6c42016e7e450e0fe2050ff321d4bd8a60c7