CHROMIUM: arm64: dts: rk3399: i2s: switch BCLK to GPIO
We discoverd that the state of BCLK on, LRCLK off and SD_MODE on
may cause the speaker melting issue. Removing LRCLK while BCLK
is present can cause unexpected output behavior including a large
DC output voltage as described in the Max98357a datasheet.

In order to:
  1. prevent BCLK from turning on by other component.
  2. keep BCLK and LRCLK being present at the same time

This patch adjusts the device tree to allow BCLK to switch
to GPIO func before LRCLK output, and switch back during
LRCLK is output.

BUG=b:181795671
BUG=b:187142184
TEST=1. Check the BCLK is turned on only 0.4 ms earlier than LRCK.
     2. Check the BCLK is turned off before LRCK.

Signed-off-by: Judy Hsiao <judyhsiao@chromium.org>
Change-Id: Ibfc40d8cdec8689e163d5f83a970f896dfbe0b2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2873308