BACKPORT: FROMGIT: ath10k: validate first subframe of amsdu before processing the list
In certain scenarios a normal MSDU can be received as an A-MSDU when
the A-MSDU present bit of a QoS header gets flipped during reception.
Since this bit is unauthenticated, the hardware crypto engine can pass
the frame to the driver without any error indication.

This could result in processing unintended subframes collected in the
A-MSDU list. Hence, validate A-MSDU list by checking if the first frame
has a valid subframe header.

Comparing the non-aggregated MSDU and an A-MSDU, the fields of the first
subframe DA matches the LLC/SNAP header fields of a normal MSDU.
In order to avoid processing such frames, add a validation to
filter such A-MSDU frames where the first subframe header DA matches
with the LLC/SNAP header pattern.

Tested-on: QCA9984 hw1.0 PCI 10.4-3.10-00047

Cc: stable@vger.kernel.org
Signed-off-by: Sriram R <srirrama@codeaurora.org>
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Link: https://lore.kernel.org/r/20210511200110.e6f5eb7b9847.I38a77ae26096862527a5eab73caebd7346af8b66@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>

(cherry picked from commit 62a8ff67eba52dae9b107e1fb8827054ed00a265
 git://git.kernel.org/pub/scm/linux/kernel/git/jberg/mac80211.git master)

Conflicts:
   drivers/net/wireless/ar10k/ath/ath10k/htt_rx.c
   [Ported htt_rx from ath10k to ar10k; decap_align_bytes is not
   supported on 4.4 so dropped int bytes_aligned = ar->hw_params.decap_align_bytes]

BUG=b:187853103, b:187853456
TEST=Verified wifi_matfunc/perf

Signed-off-by: Wen Gong <wgong@codeaurora.org>
Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: Ie4068ba26b751c38ebe0b7b0e4a770652d2749c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2891962
Reviewed-by: Brian Norris <briannorris@chromium.org>