BACKPORT: FROMGIT: ath10k: add CCMP PN replay protection for fragmented frames for PCIe
PN replay check for not fragmented frames is finished in the firmware,
but this was not done for fragmented frames when ath10k is used with
QCA6174/QCA6377 PCIe. mac80211 has the function
ieee80211_rx_h_defragment() for PN replay check for fragmented frames,
but this does not get checked with QCA6174 due to the
ieee80211_has_protected() condition not matching the cleared Protected
bit case.

Validate the PN of received fragmented frames within ath10k when CCMP is
used and drop the fragment if the PN is not correct (incremented by
exactly one from the previous fragment). This applies only for
QCA6174/QCA6377 PCIe.

Tested-on: QCA6174 hw3.2 PCI WLAN.RM.4.4.1-00110-QCARMSWP-1

Cc: stable@vger.kernel.org
Signed-off-by: Wen Gong <wgong@codeaurora.org>
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Link: https://lore.kernel.org/r/20210511200110.9ba2664866a4.I756e47b67e210dba69966d989c4711ffc02dc6bc@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>

(cherry picked from commit a1166b2653db2f3de7338b9fb8a0f6e924b904ee
 git://git.kernel.org/pub/scm/linux/kernel/git/jberg/mac80211.git master)

Conflicts:
   drivers/net/wireless/ar10k/ath/ath10k/htt_rx.c
   [Port from ath10k to ar10k; missing fill_crypt_header on 4.4
   and ieee80211_get_tid()]

BUG=b:187853103, b:187853456
TEST=Verified wifi_matfunc/perf

Signed-off-by: Wen Gong <wgong@codeaurora.org>
Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I8db035c1c03e8e0a8decf5d325de529267ac9889
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2891958
Reviewed-by: Brian Norris <briannorris@chromium.org>