CHROMIUM: Bluetooth: set quality report callback for Intel
This patch sets up set_quality_report callback for Intel to
set and reset the debug features.

(This CHROMIUM patch will be reverted by an UPSTREAM patch later.)

BUG=b:170175384
TEST=On a platform with Intel Belutooth controller, observe the
messages in the system logs.
when the quality report feature is enabled by a user process:
  Bluetooth: set_exp_feature() quality report 1 changed 1
  Bluetooth: btintel_set_debug_features() hci0: trace_enable 2 mask 127

when the vs debug event is disalbed by a user process:
  Bluetooth: set_exp_feature() quality report 0 changed 1
  Bluetooth: btintel_set_debug_features() hci0: trace_enable 0 mask 0

Change-Id: Icd1d1345b5970d3a2f8da23a10578e051e7e43e4
Signed-off-by: Joseph Hwang <josephsih@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2949069
Reviewed-by: Yun-Hao Chung <howardchung@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>