BACKPORT: FROMGIT: mac80211: extend protection against mixed key and fragment cache attacks
For some chips/drivers, e.g., QCA6174 with ath10k, the decryption is
done by the hardware, and the Protected bit in the Frame Control field
is cleared in the lower level driver before the frame is passed to
mac80211. In such cases, the condition for ieee80211_has_protected() is
not met in ieee80211_rx_h_defragment() of mac80211 and the new security
validation steps are not executed.

Extend mac80211 to cover the case where the Protected bit has been
cleared, but the frame is indicated as having been decrypted by the
hardware. This extends protection against mixed key and fragment cache
attack for additional drivers/chips. This fixes CVE-2020-24586 and
CVE-2020-24587 for such cases.

Tested-on: QCA6174 hw3.2 PCI WLAN.RM.4.4.1-00110-QCARMSWP-1

Cc: stable@vger.kernel.org
Signed-off-by: Wen Gong <wgong@codeaurora.org>
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Link: https://lore.kernel.org/r/20210511200110.037aa5ca0390.I7bb888e2965a0db02a67075fcb5deb50eb7408aa@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>

(cherry picked from commit 3edc6b0d6c061a70d8ca3c3c72eb1f58ce29bfb1
 git://git.kernel.org/pub/scm/linux/kernel/git/jberg/mac80211.git master)

Conflicts:
   drivers/net/wireless/ar10k/mac80211/rx.c
   [Ported from ath10k to ar10k]

BUG=b:187853103, b:187853456
TEST=Verified wifi_matfunc/perf

Signed-off-by: Wen Gong <wgong@codeaurora.org>
Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I46b4d38aac4cc3cf545bcad9a32e944157b0a928
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2891956
Reviewed-by: Brian Norris <briannorris@chromium.org>