CHROMIUM: iwl7000: mac8021: Allow probe request injection on passive non radar channels
Based on section 11.1.4.3.8 (Non-scanning probe request transmission)
in Draft P802.11REVmd_D5.0 transmission of a probe request to an AP
known to be in range is allowed. Thus, allow probe request injection on
monitor interface in case the target AP is found in range and the
channel is not a radar channel.

Change-Id: Iab8e73126dfe75a5448048ce118550bebda200b8
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
iwl7000-tree: 9e49abd17f015d11c08a95c3bcd1661124ec60a1