CHROMIUM: iwl7000: mvm: Don't convert rate_n_flags of last_phy_info
The field last_phy_info of the iwl_mvm structure was deprecated
and therefore contains the old rate only. We used to convert it,
but on the other hand we used the old rate definitions.
Fix it by not converting.

Change-Id: I77a0a2b6c39c01f8381aa3c79f52c8c9571219cc
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
iwl7000-tree: 0250f85e2b15d36ef61f47434f0f854d782c04a9